Add some utility functions
